refactor: split enemies.rs into themed submodules
Break the 2945-line enemies.rs into an enemies/ module directory: tables
(class data), sprite_bank (CHR model), class_modes (mode/pool resolution),
picking (selection mechanics), segments (HB wild segments), injection
(wild injection), and tests. mod.rs keeps the orchestration entry points
and the main object-data walker.

Also fix a latent fragility the split exposed: the walker classifies
find_class_pool's result with std::ptr::eq against PIRANHAS_WILD /
PIRANHASC_WILD. A const's promoted backing array has no guaranteed unique
address and gets duplicated once the producer and comparison live in
different codegen units, silently breaking the identity check (and
desyncing the RNG stream). Making the two pools `static` guarantees one
address and keeps output identical.

Verified byte-for-byte identical ROM output across seeds and the full
flag matrix; all tests pass, clippy clean.
